refactor connect module out of agent/agent.rb
Moved the connect module into its own file from lib/new_relic/agent/agent.rb to reduce
the number of modules living inside of that file
