+++ test/multiverse/suites/rake/rake_test.rb
@@ -44,7 +44,7 @@ if ::NewRelic::Agent::Instrumentation::Rake.should_install? &&
         end
 
         refute_any_rake_metrics
-        assert_includes @output, "ERROR : NewRelic::Agent::Agent::InstanceMethods::Connect::WaitOnConnectTimeout: Agent was unable to connect"
+        assert_includes @output, "ERROR : NewRelic::Agent::Agent::Connect::WaitOnConnectTimeout: Agent was unable to connect"
       end
 
       def test_records_transaction_metrics
